Extract PushProcessEnd step for pipeline
