+++ repairnator/repairnator-core/src/main/java/fr/inria/spirals/repairnator/states/PushState.java
@@ -8,5 +8,6 @@ public enum PushState {
     REPO_INITIALIZED, REPO_NOT_INITIALIZED,
     REPAIR_INFO_COMMITTED, REPAIR_INFO_NOT_COMMITTED,
     PATCH_COMMITTED, PATCH_NOT_COMMITTED,
-    REPO_PUSHED, REPO_NOT_PUSHED
+    REPO_PUSHED, REPO_NOT_PUSHED,
+    END_PUSHED
 }
