+++ repairnator/repairnator-pipeline/src/main/java/fr/inria/spirals/repairnator/process/inspectors/ProjectInspector.java
@@ -13,6 +13,7 @@ import fr.inria.spirals.repairnator.process.step.paths.ComputeTestDir;
 import fr.inria.spirals.repairnator.process.step.push.InitRepoToPush;
 import fr.inria.spirals.repairnator.process.step.push.PushIncriminatedBuild;
 import fr.inria.spirals.repairnator.process.step.push.CommitPatch;
+import fr.inria.spirals.repairnator.process.step.push.PushProcessEnd;
 import fr.inria.spirals.repairnator.process.step.repair.AbstractRepairStep;
 import fr.inria.spirals.repairnator.states.ScannedBuildStatus;
 import fr.inria.spirals.repairnator.notifier.AbstractNotifier;
@@ -52,6 +53,8 @@ public class ProjectInspector {
 
     private CheckoutType checkoutType;
 
+    private AbstractStep finalStep;
+
     public ProjectInspector(BuildToBeInspected buildToBeInspected, String workspace, List<AbstractDataSerializer> serializers, List<AbstractNotifier> notifiers) {
         this.buildToBeInspected = buildToBeInspected;
 
@@ -65,6 +68,8 @@ public class ProjectInspector {
         this.notifiers = notifiers;
         this.checkoutType = CheckoutType.NO_CHECKOUT;
         this.initMetricsValue();
+
+        this.finalStep = new PushProcessEnd(this);
     }
 
     private void initMetricsValue() {
@@ -208,5 +213,9 @@ public class ProjectInspector {
 
     public void setPatchNotifier(PatchNotifier patchNotifier) {
         this.patchNotifier = patchNotifier;
+    }
+
+    public AbstractStep getFinalStep() {
+        return finalStep;
     }
 }
